upgrade diversifier to a uint64_t

+++ src/dart_interface.cpp
@@ -65,14 +65,15 @@ const char* createIncomingViewKey(const char* keyData, int index) {
  * FFI-friendly wrapper for spark:getAddress.
  */
 EXPORT_DART
-const char* getAddress(const char* keyDataHex, int index, int diversifier, int isTestNet) {
+const char* getAddress(const char* keyDataHex, int index, uint64_t diversifier, int isTestNet) {
     try {
         // Use the hex string directly to create the SpendKey.
         spark::SpendKey spendKey = createSpendKeyFromData(keyDataHex, index);
 
         spark::FullViewKey fullViewKey(spendKey);
         spark::IncomingViewKey incomingViewKey(fullViewKey);
-        spark::Address address(incomingViewKey, static_cast<uint64_t>(diversifier));
+        // spark::Address address(incomingViewKey, static_cast<uint64_t>(diversifier));
+        spark::Address address(incomingViewKey, diversifier);
 
         // Encode the Address object into a string using the appropriate network.
         std::string encodedAddress = address.encode(isTestNet ? spark::ADDRESS_NETWORK_TESTNET : spark::ADDRESS_NETWORK_MAINNET);
